changes for snyk auth

+++ build_thirdparty.sh
@@ -40,6 +40,10 @@ set -x
 # shellcheck disable=SC2086
 python3 "$YB_THIRDPARTY_DIR/python/yugabyte_db_thirdparty/yb_build_thirdparty_main.py" "$@"
 
+rc=$?
+set +x
+
+
 if [[ $* == *--snyk* ]]; then
   if [[ $OSTYPE == linux* ]]; then
     curl https://static.snyk.io/cli/latest/snyk-linux -o snyk
@@ -48,6 +52,12 @@ if [[ $* == *--snyk* ]]; then
   fi
   chmod +x ./snyk
   homedir="${BASH_SOURCE[0]%/*}"
+  ./snyk auth "${SNYK_TOKEN}"
+  rc2=$?
+  if [[ $rc2 -ne 0 ]]; then
+    exit $rc
+  fi
   ./snyk test "$homedir/src" --unmanaged
   rm -f ./snyk
 fi
+
